+++ packages/experimental/Menu/src/MenuTrigger/MenuTrigger.tsx
@@ -31,19 +31,18 @@ MenuTrigger.displayName = menuTriggerName;
 
 const getRevisedProps = memoize(getRevisedPropsWorker);
 function getRevisedPropsWorker(state: MenuTriggerState, props: any): MenuTriggerChildProps {
-  const revisedProps = { ...state.props };
-
+  const revisedProps = state.props;
   if (props.accessibilityState) {
-    revisedProps.accessibilityState = { ...state.props.accessibilityState, ...props.accessibilityState };
+    revisedProps.accessibilityState = { ...revisedProps.accessibilityState, ...props.accessibilityState };
   }
 
   if (props.accessibilityActions) {
-    revisedProps.accessibilityActions = { ...state.props.accessibilityActions, ...props.accessibilityActions };
+    revisedProps.accessibilityActions = { ...revisedProps.accessibilityActions, ...props.accessibilityActions };
   }
 
   if (props.onAccessibilityAction) {
     revisedProps.onAccessibilityAction = (e: AccessibilityActionEvent) => {
-      state.props.onAccessibilityAction(e);
+      revisedProps.onAccessibilityAction(e);
       props.onAccessibilityAction(e);
     };
   }