+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Fri, 03 Jun 2022 19:33:31 GMT and should not be manually modified.
+This log was last generated on Mon, 06 Jun 2022 08:34:26 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.63.3
+
+Mon, 06 Jun 2022 08:34:26 GMT
+
+### Patches
+
+- Fixed SVG icon size and color for the Avatar (v.kozova13@gmail.com)
+- Bump @fluentui-react-native/experimental-avatar to v0.14.36
+
+## 0.63.2
+
+Fri, 03 Jun 2022 21:01:53 GMT
+
+### Patches
+
+- Add a test for infinite submenus (sanajmi@microsoft.com)
 
 ## 0.63.1